Move SearchTools to helpers since it can be a shared component, update dnsdb to use search tools for data export, modify how pagination works to support search tools, remove unnecessary text_handler files
